add user id auth to button on show

+++ app/Http/Controllers/PostsController.php
@@ -78,7 +78,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        return view('posts.edit')->with('post', $post);
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthized Page');
+        }
+        return view('posts.edit')->with('post', $post);        
     }
 
     /**
@@ -109,7 +113,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
+        
     }
 }
